refactor(schemabuilders/list): allow slice entities, do not make list if already slice
