+++ utils/pluralize.go
@@ -1,0 +1,7 @@
+package utils
+
+import (
+	p "github.com/gertd/go-pluralize"
+)
+
+var Pluralize = p.NewClient().Plural
